% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomag_map.R
\name{geomag_map}
\alias{geomag_map}
\title{Compute Magnetic Likelihood Maps}
\usage{
geomag_map(
  tag,
  compute_known = FALSE,
  sd_e_f = 0.007,
  sd_e_i = 5,
  sd_m_f = 0.01,
  sd_m_i = 3.2,
  ref_map = geomag_map_ref(tag),
  quiet = FALSE
)
}
\arguments{
\item{tag}{A \href{https://raphaelnussbaumer.com/GeoPressureR/reference/tag_create.html}{GeoPressureR tag object }.}

\item{compute_known}{Logical. If TRUE, computes likelihood maps for known stationary periods;
if FALSE, fixes the likelihood at the known location.}

\item{sd_e_f}{Numeric. Standard deviation of observation noise of intensity error (single value
or one per stap).}

\item{sd_e_i}{Numeric. Standard deviation of observation noise of inclination error (single
value or one per stap).}

\item{sd_m_f}{Numeric. Standard deviation of stationary-period-specific noise of intensity error
(single value or per stap).}

\item{sd_m_i}{Numeric. Standard deviation of stationary-period-specific noise of inclination
error (single value or per stap).}

\item{ref_map}{Raster stack or list. Reference magnetic maps (intensity and inclination)
computed by default using \code{geomag_map_ref()}. Provide it if you've already computed it to save
computational time.}

\item{quiet}{Logical. If TRUE, suppresses progress messages.}
}
\value{
A \href{https://raphaelnussbaumer.com/GeoPressureR/reference/tag_create.html}{GeoPressureR tag object } with likelihood maps
added as:
\itemize{
\item \code{tag$map_magnetic_intensity}
\item \code{tag$map_magnetic_inclination}
\item \code{tag$map_magnetic}
and updated parameters in \code{tag$param}.
}
}
\description{
Estimates a likelihood map for each stationary period based on observed magnetic data
(intensity and inclination) compared to the expected values from the World Magnetic Model (WMM).
}
\examples{
library(GeoPressureR)
withr::with_dir(system.file("extdata", package = "GeoMag"), {
  tag <- tag_create("14DM", quiet = TRUE)
  tag <- tag_label(tag, quiet = TRUE)
  tag <- tag_set_map(tag,
    extent = c(-18, 23, 0, 50),
    scale = 2,
    known = data.frame(
      stap_id = c(1, -1),
      known_lon = 7.27,
      known_lat = 46.19
    )
  )
  tag <- geomag_calib(tag, quiet = TRUE)
  tag <- geomag_map(tag, quiet = TRUE)
  plot(tag, type = "map_magnetic_intensity")
  plot(tag, type = "map_magnetic_inclination")
})
}
